mox-support: improve number prefix for Lighttpd configuration file
Due to recent change in Lighttpd, when it prefixes the configuration
files, we decided to do the same for all our lighttpd configuration
files, but reForis redirect can not start earlier than reForis itself and other
dependencies like Foris WS, etc.

The numbering for reForis redirect with recent changes was invalid,
because reForis does not want to start on Turris Shield, thus fixing it
to having the same prefix as reForis or later.

Fixes: hash ("treewide: use number prefixes for lighttpd configuration
files")