turris-diagnostics: add dependency lighttpd-mod-fastcgi for web packages
In Turris Diagnostics 16.0, there were added two webs diagnostics and
snapshots, which required lighttpd-mod-fastcgi and lighttpd-mod-cgi [1].

There was added lighttpd-mod-fastcgi, but we forgot to add
lighttpd-mod-cgi. The later dependency is present in Turris WebApps,
but Turris WebApps are not available on Turris Shield device and that's
why we did not notice that.

[1] https://gitlab.nic.cz/turris/diagnostics/-/blob/ffb7d19ec9fe7468a790be96fdf46989a9dcef03/web/lighttpd.conf#L3

Fixes: https://gitlab.nic.cz/turris/os/packages/-/issues/881
Suggested-by: Martin Matějek <martin.matejek@nic.cz>