sentinel-fwlogs: revert several commits in master branch
This ensures, well... to be correct.
This only revert several commits, which are causing segfaults and resulting in fwlogs in an endless crash loop. Mirek found out that this
the issue is not in the commits itself but in the 3rd-party library.
However, several users are complaining that they have ugly false check,
which they have in reForis, thus let's do this way, which I am not proud of.

Reference: https://gitlab.nic.cz/turris/sentinel/fwlogs/-/issues/9