nextcloud: Migrate to upstream updating solution
Nextcloud has it's own updater. With this release, we are switching from
supplying packages for Nextcloud and maintaining those to a setup where
we provide initial installation and then we let Nextcloud keep itself up
to date all by itself (with the help from cron).

Before setup, only placeholder webpage that redirects to reForis is
used. Migration from previous version should be covered.

One issues that this setup solves is downgrades that are not supported,
but might happen when switching between branches.