lighttpd-https-cert: transition from uci-defaults to init
After the initial boot, uci-defaults creates self-signed certificate for
lighttpd. This certificate will never be renewed and is set to expire in 20
years. The init script manages expiration date, and if the certificate
is about to expire in 4 days or has already expired, it will be renewed upon the
next reboot.