sentinel-certgen: Force distutils-based python package builds
In OpenWRT main / 23.05 the new build process chokes on
the OpenWRT package names differing from the python package
names. Force the use of the previous build process for now.

Alternative solutions include providing a PYPI_SOURCE_NAME that
contains the python package name, or renaming the OpenWRT packages
to have the same name as the python package.

Signed-off-by: Magnus Kessler <Magnus.Kessler@gmx.net>