foris-controller-storage-module: Disable logging to /srv by default
If an external storage is mounted, logging to /srv/log is enabled by
default. As reported, it may have unexpected and unwanted consequences
(e.g. consumption of free space on the storage or slowing down due to
eventual poor performance of the given storage).  Thus we should keep
the logging disabled by default. It can be explicitly enabled if
required.