modem-manager-autosetup: Fix jq filter for apn without 'type' attribute
The initial jq filter used to extract the APN info from apns-conf.json
fails with:

``` text
jq: error (...): null (null) and string ("default") cannot have their containment checked
```
when one of the apn objects selected by mcc/mnc doesn't have a
'type' attribute (as is the case for mcc=262, mnc=07). As a result,
otherwise matching APN objects will not be considered.

This commit fixes the jq filter by checking for the presence of
the 'type' attribute.