modem-manager-autosetup: Use optional mvno_match_data for APN selection
In case that there are multiple APN definitions for the given
combination of 'mnc', 'nmc' (and 'type') and an 'mvno_match_data'
attribute of type 'imsi' exists, the 'mvno_match_data' can be used to
select the correct APN from a set of otherwise equivalent ones.

This commit adds an additional step which uses 'mvno_match_data' to
improve the APN selection process.