sentinel-firewall: Assume clean operation
Let's assume that nobody is interfering and we are running just in one
instance. Therefore we can cleanup everything (without expensively
checking for whatever exists) and set-up all the bits and pieces we will
need upfront during the cleanup to again avoid painfully expensive
checking whether everything is already set-up.