rescue-image: Latest Omnia NG related fixes
We no longer need to build the U-Boot environment, resetting it should
suffice. We have a only one and fixed rescue interface and that is eth3.
And as we are near production, we no longer include the test keys.
Also drop runtime dependencies as all we have are build dependencies.