suricata-bin: make libunwind dependency as conditional
In OpenWrt, libunwind is not available for powerpc [1],
unless someone decided to look at it and make it work
together with libucontext [2] according to libunwind
documentation [3].

Suricata works also without libunwind. Libunwind
is used for stacktrace support [4].

This commit makes libunwind a conditional dependency,
so if it's available (e.g., on x86_64, aarch64), Suricata
will be compiled with stacktrace support; otherwise it
will work without it.

[1] https://github.com/openwrt/openwrt/blob/c3c75d0e685b90fdcb9f1cf79a6e2c6949cce9ac/package/libs/libunwind/Makefile#L35
[2] https://redirect.github.com/openwrt/openwrt/commit/fca63d04b5075723d001eff788973edd88bcb636#commitcomment-167037367
[3] https://github.com/libunwind/libunwind?tab=readme-ov-file#libc-requirements
[4] https://docs.suricata.io/en/suricata-8.0.1/configuration/suricata-yaml.html#stacktrace

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>